refactor(rack): reine Draft-Helfer aus RackBuilderDialog auslagern (#468)
RackBuilderDialog.tsx 1867 → 1692 Zeilen. Konstanten (Rack-Maße, Row-
Heights, DRAFT_KEY) + reine Draft-Transformer (parseUnits, toPlacement,
normalizeDraft, draftFromPreset, formatRackUnits, sanitizeTemplatePorts)
wandern nach rackBuilderHelpers.ts.

Verhaltensneutral: identischer Code, nur verschoben + importiert.
tsc 0, eslint 0, build 0.

Refs #468
